Parsing Make and JX unified entry point
