+++ makeflow/src/parser_make.h
@@ -1,0 +1,12 @@
+/*
+Copyright (C) 2008- The University of Notre Dame
+This software is distributed under the GNU General Public License.
+See the file COPYING for details.
+*/
+
+#ifndef PARSER_MAKE_H
+#define PARSER_MAKE_H
+
+int dag_parse_make(struct dag *d, FILE * dag_stream);
+
+#endif
